requirements-dev.txt
mypy
python>=3.10
pytorch >=1.12,<2.1
ruff
yapf
